Extract LocationButton into a separate module
